BUG(11): removed warning.
The flag VTK_EXPORT was defined as dllimport when building vtkTkRenderWidget.
This needed to be dllexport.  However, VTK_EXPORT needed to be defined as
dllimport for proper inclusion of other vtk functions.  I added a VTK_TK_EXPORT
local to vtkTkRenderWidget that defaults to dllexport on the PC and to nothing
on a unix machine.