Modified vtkLight to include a transformation matrix to make it possible to automatically track the camera cleanly (among other things). vtkOpenGLLight implements the transformation matrix functionality by handing the transpose of the matrix to OpenGL.
Added support for a CameraLight coordinate system in vtkCamera.  In this
coordinate system, the camera is always at (0, 0, 1), looking at (0, 0, 0),
with focal distance of 1, and with up being (0, 1, 0).  The camera
automatically computes this matrix.

These changes are backwards compatible with existing code.  Additional
changes to vtkRenderer and the interactors that use the transformation
functionality will follow, once I figure out how to make the changes
as backwards-compatible as possible.