ERR: Reported by John Tourtellott <johnt@raytechcomputing.com> At 09:36 AM 10/31/00 -0500, John Tourtellott wrote: I believe that the vtkOBJReader class never reads vertex normals from an input file, due to a bug in vtkOBJReader.cxx at lines 208-209:
    208   blank = (char *) strchr (line, (int) ' ');
    209   slash = (char *) strchr (line, (int) '/');
    210   if (blank && slash && (slash < blank))
    211      {

This code is in the middle of parsing a vertex/texture/normal string for the next  index into the array of normals. Because the current code parses from the beginning of the 'line' variable, the resulting 'blank' value is always less than the 'slash' value and, as a result, the subsequent if statement (line 210) always evaluates to false. Instead, the parsing should be done with respect to the current pointer position, i.e., the correct code is:

    208   blank = (char *) strchr (ptr, (int) ' ');
                                   ^^^
    209   slash = (char *) strchr (ptr, (int) '/');
                                   ^^^