Merging of my changes with those made by Kitware, including:
Rewrote Charles' rewrite to use less memory.  Still using his
intersection operations on vtkDataSetAttributes::FieldList objects
so that the order of field arrays will be correct when data sets
are merged, but you now need to specify an upper bound on the
total number data sets before you start.

I rewrote the merging of cell data to detect when the input
is a vtkUnstructuredGrid and merge more efficiently in that
case.  It's 10% faster.

I fixed all the curly braces to match Kitware's style standards.

I fixed all cell/point Ids to be vtkIdTypes.  Kitware had
already fixed many of these.