ENH: a) the vtkDetermineTkResources.cmake CMake code was not *just* determining the Tk resource file, it was also configuring some other files and variables that are *only* required to build the vtk.exe shell, etc, which is a bit confusing since *other* projects like PV or VV are calling including the CMake file, without knowing that it does much more than it should. Not a problem, but I cleaned it up by leaving only the code that really finds the resource file inside it, and move the rest where it is only needed, i.e. in Wrapping/Tcl/CMakeLists.txt
b) the resulting variable VTK_TK_RESOURCE_FILE, points to the tk.rc file in the Tk resources directory. This is not a big deal, but what matters here is the *directory*, not the file, because *all* the files in that directory are required (tk.rc includes all of them), so pretty much all the projects that were using VTK_TK_RESOURCE_FILE were to perform a GET_FILENAME_COMPONENT afterwards, etc. I simplified the process by exporting VTK_TK_RESOURCES_DIR instead of VTK_TK_RESOURCE_FILE.

c) a IF (VTK_TK_RESOURCE_FILE) test was performed in Wrapping/Tcl/CMakeLists.txt to see if we could actually create a vtk.rc file, i.e. a resource file for the vtk.exe shell that would include both tk.rc and some other stuff. That test is wrong (too restrictive), because tk.rc is only needed if Tcl/Tk is static, and there is already a test for that in vtk.rc. If Tcl/Tk is not static, and the resource file was not found for some reason, that would still work now.

d) the vtkDetermineTkResources.cmake macro was called from Wrapping/Tcl, which only happens if you go inside that dir. This is too restrictive since you *still* need that Tk resource file/dir to build C++ applications based on KWWidgets+VTK+Tcl/Tk-Static (for example), even if you did not build a vtk.exe shell (PV or VV do not wrap Tcl, so they do not build the shell and would not export the resource properly I think). So it is now called from the main CMakeLists.txt, inside the VTK_USE_TK test, right after we include the Tcl/Tk macros.