ENH: BUG: 1) The summation interface has been removed from the gui. I replaced it with "Select mass arrays". the number of mass arrays have to be the same as the number of material fraction arrays selected, and they are paired, the first mass array is used with the first material, the second with the second and so on. In addition to mass I am now computing moments and these are stored in a 4 component tuple (Myz, Mxz, Mxy, m). So that one can now get center of mass for each fragment if desired.
2)
The filter's output is now "multi-block -> multi-piece -> poly data". Where: each piece is a material, and each poly data is a fragment of that material. The attributes, Id, Volume, Mass, Moments, & Weighted Averages, are now single tuple field data in each fragment, as well as copied into the fragments point data. In the future future when PV has the capability  to color a surface by dataset attribute we can hopefully stop copying the attributes into the point data. This fixes the output so that applying additional filters downstream works.

3)
The source arrays which are selected for volume weighted averaging are copied into the point data so that one can color fragments by the original data. This makes for nicer, and  more informative graphics than the average which is constant on the fragment. The averages are also in the point data so you have the choice.

4)
As a side effect of running the filter a csv text file can be generated with all the fragment attributes. The gui now has a check box to enable writing of this file and a text input to specify the path and file name base. Each process will write one file containing the fragments it owns per each material processed. The file name is: "/path/to/base.N.M.cthfc" where N is the process id and M is the material id. The files contain a Header, a list of attributes and their respective number of components, and the csv data. The tuples are separated by commas while the components of each tuple are separated by spaces. The file format can be deduced from the following example:

Header:{
Process id:0
Number total:2
Number local:2
Number of attributes:7
}
Attributes:{
"Id", 1
"Volume", 1
"Fragment Moments", 4
"Mass", 1
"VolumeWeightedAverage-Mass (g) - 1", 1
"VolumeWeightedAverage-Velocity", 3
"Summation-Velocity", 3
}
Attribute Data:{
0, 9.5, 12 0 0 60.8, 60.8, 0.8, 0.5 0.5 0, 38 38 0
1, 1, 12.8 0 0 6.4, 6.4, 0.8, 0.5 0.5 0, 4 4 0
}