ENH: Change the z-shift method in vtkOpenGLCoincidentTopologyResolutionPainter to shift lines in front of polygons rather than normal polygons in front of triangle strips (the latter of which makes no sense).
ENH: Apply a Z-shift before the perspective transformation rather than after it (which is what the glDepthRange effectively did).  This seems to provide a bit better results.

ENH: Use a Z-shift to resolve coincident topology in ParaView rather than a polygon output.  The polygon output uses the "z-slope" of polygons to determine the amount to shift, and as the z-slope approachs infinity as polgons approach perpendicular to the view plane, the resulting z-shift caused primitives to pop in front of and behind others.