ENH: Refactoring the time management in ParaView a bit to move the time step management logic to the server manager. vtkSMTimeKeeperProxy now provides API to add time sources from which it obtains information about time steps and on it's own keeps the "TimestepValues" and "TimeRange" properties updated. (Earlier this was done by pqTimeKeeper).
Animation Scene has a new mode UseCustomEndTimes. Which when set results in
using the end times specified by the user. Otherwise the time range in obtained
from the time keeper and hence the animation can be scaled automatically if the
data time changes when the state file is loaded etc.

Add "precision" attribute for vtkSMDoubleVectorProperty which can be used to
fine tune the precision at which the double values are saved in state files.
Time and AnimationTime are now saved with 16 decimal precision.

When state file is loaded, UpdatePipelineInformation() is called on all source
proxies. This ensures that all information properties get updated correctly when
state files are loaded. This ensures that TimestepValues etc. are re-read from
the reader when state files are loaded as well.