The following all applies to FindCell.
PERF: Check to see if the given coordinates are outside the bounds and fail
quickly if they are.

ENH: If we fail to find the cell from the given starting point and from
a cell collected to the closest point to the coordinates, find all points
coincident to the first one and trace from all neighboring cells to those.
This can help find cells when their are duplicate points.

PERF: As we trace through cells, keep track of all cells we have visited
and terminate walk whenver we hit one.

STYLE: Restructure the FindCell method to make logical subcalls as we
try different techniques.  IMHO this is easier to read.  It even reduces
the lines of code after adding the other features.