ENH: Added ability to force a render strategy in the device, improved test.
Added functions in OpenGLContextDevice2D to force a strategy, this enabled
the TestDiagram test to be improved and always use the freetype strategy.
Also made some improvements to save and restore GL state, and handle the case
where Begin is called more than once, although this should possibly result in
an error as Begin calls should be paired with End calls.