ENH: Enable floating point warnings on all Linux architectures.
Spoke with Francois about enabling floating point warnings on all Linux
architectures, as there are no continuous 64 bit Linux builds it will need to be
tested on the Nightly builds. Works locally, but it is not ideal as it calls
exit. Currently tests are segfaulting, or exiting on Windows. On x86_64 Linux
machines no errors were raised before.