COMP: Adding back in a patched FindPythonLibs module to deal with Mac frameworks.
Thanks to David Gobbi for reporting an issue with Python frameworks when using
the CMake find module. This module is a modified form of the CMake module, and
was successfully tested by David Gobbi. If there are no further issues these
changes should be merged upstream to CMake for their next release.