ENH: Adding my first pass at an experimental vtkContextView based XY chart.
Added a new view to ParaView based on the new vtkContextView work I have been
doing in VTK.  This code is experimental, and currently only works with the
line plot filter. I am actively working on this code, the API may change a
little and the interface and view options will certainly be improved.

The ContextView proxies are intended to remain very general so that further
context view based charts can be added to ParaView more easily. Feedback would
be appreciated, I am currently working on adding the proxies to choose the
columns to be plotted and to customize the plots (color, thickness, symbols).