COMP: Fixes for GCC visibility support - implicit template instantiation issues.
Thanks to Brad King for helping me to track this down! The vtkImageIterator
was being implicitly instantiated in vtkImageProgressIterator, causing the used
symbols to be hidden. Do not include vtkImageIterator.txx in the progress
iterator implementation to avoid this issue. Should work everywhere now - correct
fix was to remove the .txx include!